Bug fixes per Sonarqube.

+++ src/com/wasteofplastic/askyblock/generators/ChunkGeneratorWorld.java
@@ -144,7 +144,7 @@ public class ChunkGeneratorWorld extends ChunkGenerator {
                     }
                     // Next three layers are a mix of netherrack and air
                     for (int y = 5; y < 8; y++) {
-                        double r = gen.noise(x, maxHeight - y, z, 0.5, 0.5);
+                        double r = gen.noise(x, (double)maxHeight - y, z, 0.5, 0.5);
                         if (r > 0D) {
                             setBlock(result, x, (maxHeight - y), z, (byte) Material.NETHERRACK.getId());
                         } else {
@@ -152,7 +152,7 @@ public class ChunkGeneratorWorld extends ChunkGenerator {
                         }
                     }
                     // Layer 8 may be glowstone
-                    double r = gen.noise(x, maxHeight - 8, z, random.nextFloat(), random.nextFloat());
+                    double r = gen.noise(x, (double)maxHeight - 8, z, random.nextFloat(), random.nextFloat());
                     if (r > 0.5D) {
                         // Have blobs of glowstone
                         switch (random.nextInt(4)) {
